Improve error handling for Fleep HTTP responses
The Fleep API sometimes responds with HTTP 5XX status codes
and/or with invalid JSON. This causes the adapter to hang.

This commit moves JSON decoding of the response body
after the HTTP status code check and surrounds it with
a try/catch statement. This should avoid JSON decoding
exceptions in the case of invalid responses.

Closes #10